Move templated method to detail
