+++ src/perception/DetectedObject.cpp
@@ -73,22 +73,5 @@ void DetectedObject::setMetaSkeleton(
   mMetaSkeleton = metaSkeleton;
 }
 
-//==============================================================================
-template <typename T>
-T DetectedObject::getInfoByKey(const std::string& key) const
-{
-  T value;
-  try
-  {
-    value = mYamlNode[key].as<T>();
-  }
-  catch (const YAML::ParserException& ex)
-  {
-    throw std::runtime_error(
-        "[DetectedObject] Error in converting [" + key + "] field");
-  }
-  return value;
-}
-
 } // namespace perception
 } // namespace aikido
